[fix] Get tempo placement based on duration
Using a modified version of Dynamic Programming solve of the Coin Change problem, the program can determine the optimal combinatios of durations to note types to get a specific amount of duration.
We use that to insert the tempo element based not on the element index, but on the duration.